ai_atl_2025-master/requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
